+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- Example to test openDir on /home/web/tmp folder.
+### Changed
+### Fixed
+
+
+## [210220] - 20210220
+### Added
+- Example to confirm openDir on /home/web/tmp folder fails.
+- Example to confirm outbound network connection fails.
 ### Changed
 - Prevent Zig openDir on /home/web/tmp folder.
 - Title lookup from Zig source files - check test.zig after main.zig.